@@ -1,11 +1,11 @@
 name: progress_indicator_button
 description: A simple button which can transform and show a progress indicator.
 version: 0.0.4
-author: Pascal Brosinski<pascal@brosinski.eu>
 homepage: https://github.com/PascalAC/progress_button
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.0.0"
 
 dependencies:
   flutter:
