@@ -27,8 +27,8 @@ class ProgressButton extends StatefulWidget {
   final Widget? child;
 
   ProgressButton({
-    @required this.child,
-    @required this.onPressed,
+    required this.child,
+    required this.onPressed,
     this.color = Colors.blue,
     this.strokeWidth = 2,
     this.progressIndicatorColor = Colors.white,
@@ -43,7 +43,7 @@ class ProgressButton extends StatefulWidget {
 
 class _ProgressButtonState extends State<ProgressButton>
     with TickerProviderStateMixin {
-  var _controller;
+  late AnimationController _controller;
 
   @override
   void initState() {
@@ -65,7 +65,7 @@ class _ProgressButtonState extends State<ProgressButton>
   Widget build(BuildContext context) {
     return Center(
       child: ButtonStaggerAnimation(
-        controller: _controller.view,
+        controller: _controller,
         color: widget.color,
         strokeWidth: widget.strokeWidth,
         progressIndicatorColor: widget.progressIndicatorColor,
